Minor change to retrigger blog publishing
I submitted this post with a future date of September 16 and thought it
would be visible on the Sonobuoy website on that date. Checking the
webste this morning, this post is not visible so it didn't work as
expected.

To retrigger the publishing of the post, this change just includes a
minor change that does not impact the content or style (removing an
extra blank line).

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>