Make timeouts more clear
This change has the ultimate aim of making sure that when Sonobuoy
times out, the messaging to the user is clear.

 - Makes the timeout message in the logs more clear by adding another
closer to the end of the logs (rather than in the middle of the run
before the queries take place)
 - Makes a new status for plugins `timeout` which will be reported
via `sonobuoy status` and `sonobuoy results`. This new timeout
status will not bubble up and will just apply to the specific plugins
that had not reported results by the deadline. It will bubble up as if
it were a normal failure (i.e. its parent will be marked as failed not
as timeout)
 - Avoids "cleaning up" plugins due to timeout. This logic was added
so that plugins might be sent a signal (alerting them to the timeout)
and they could report results. However, the problem is that the results
can be very misleading since they are forced via the timeout. In the
case of E2E tests, even when 100 tests are to be run, if a run is
cut off due to timeout, only a subset of those tests are reported
on and the timeout is not reported. In addition, this cleanup logic
prevent may prevent the pods from remaining when queries take place
giving incomplete debug information.

Fixes #883

Signed-off-by: John Schnake <jschnake@vmware.com>