Adds a testImage for our integration tests
Adds the following:
 - a Dockerfile to build a testImage
 - a Makefile to build the image
 - a script to generate YAML plugin definitions for some test cases
 - the generated plugin definitions from the above script
 - source code to run a test image in two modes: single file or
tarball results
 - raw data which the test image will report to Sonobuoy as its
test results

The test image is meant to be invoked with either single-file
or tar-file mode. The arguments after that are the file[s] which
to report as the plugins results.

This allows us to generate passing/failing/skipped tests and
report on all the combinations with just having some static
results files and various plugin definitions which invoke the
image in different ways.

The path to the files to report, when invoking the plugin,
should be stated as absolute paths, rooted at /resources (analogous
to how they are currently put into the resources directory).

Related to #668

Signed-off-by: John Schnake <jschnake@vmware.com>