Prints all plugins when invoking `sonobuoy results` by default
If given a tarball, there is not a way to detect which plugins
were actually run without opening it up first. This means that the
`sonobuoy results` command is less useful since you have to have
preexisting knowledge about the tarball/run. You'd also have to
manually execute the command numerous times if you want all the
summaries of the plugins.

This change:
 -  makes the default value for the `--plugin` flag the
empty string
 - when a string value is given, it will print a single report
like the existing behavior
 - when the default, empty string is used, we will list the report
for each plugin
 - adds a new file, meta/info.json which will contain the list of
plugins that were actually run so that it is easier to grab this
list (rather than more complicated logic using the other data in
the tarball)

Fixes #905

Signed-off-by: John Schnake <jschnake@vmware.com>