Default to using the golang-based conformance runner
In Kubernetes 1.16.0 a problem in the bash runner of the conformance
tests caused the plugin to fail to report results whenever a test
failure occured.

This problem does not impact the golang-based runner which was
introduced also in Kubernetes 1.16.0. By default we would like to use
this feature as a result. It is a new feature but has been shown
so far to be functional and an improvement over the bash-based
runner (e.g. like this issue in general).

Users of Sonobuoy v0.16.0 should update to a version which utilizes
this env var by default, but they can also specify
--plugin-env e2e.E2E_USE_GO_RUNNER=true to workaround the issue.

If users of new versions of Sonobuoy wish to NOT use the go runner
they can also set --plugin-env e2e.E2E_USE_GO_RUNNER to unset the value.

Fixes #910

Signed-off-by: John Schnake <jschnake@vmware.com>