Allow piping input to sonobuoy run
Often a user wants to use `sonbouoy gen` to generate some YAML and
then modify it using tools like jq or yq. We even have this workflow
documented and the user has to, end the end, pipe the data into
`kubectl apply -f`.

`sonobuoy run` is just a shorthand for `sonobuoy gen` + execute with
wait logic; it makes sense to allow loading a file or piping the data
from stdin.

This PR:
 - adds a --file, -f flag to `sonobuoy run`
 - if the value is `-` it will load from stdin, erroring if it is not
connected to a terminal.

It does not add a unit test for this since the `sonobuoy run` logic
is difficult to test in a unit test since it wants to talk to a real cluster.

We may consider adding this to integration tests but it is mostly just input
transformation and leaves little room for error.

Fixes #920

Signed-off-by: John Schnake <jschnake@vmware.com>