Sleep forever after systemd logs plugin finishes
We currently sleep for 1h after the logs are gathered but some
test runs take multiple hours which leads to this plugin getting
restarted and trying to resubmit logs. This causes confusing messages
in the logs since the original plugin pod is gone and the only one that
exists gets a 409 error from submitting duplicate results.

Fixes #970

Signed-off-by: John Schnake <jschnake@vmware.com>