Set OwnerReference on plugin TLS Secret
We previously modified plugin Pods and DaemonSets to be owned by the
main Sonobuoy pod. When running plugins, we also create a secret for the
TLS certificate used for communication between plugins and the
aggregator pod. This change adds an OwnerReference to this secret so
that it is cleaned up if the main Sonobuoy aggregator pod is deleted.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>