Use prerelease info when looking up conformance image (#978)
Upstream publishes prerelease images so there is no need to throw
that info away. It is a remnant of Heptio publishing their own
images.

In addition, if we're using that info we dont need to warn about
the existence of prerelease info and never really needed to warn
about the metadata on the version. As a result we can remove that
entire warning check/log.

Fixes #976
Fixes #977

Signed-off-by: John Schnake <jschnake@vmware.com>