Allow daemonsets to use affinity to limit nodes they run on (#990)
For daemonsets, node affinity can be set to limit/select the nodes
which the daemonset will run on. Since this is a native k8s feature
we should support it.

However, we use 'node-plugin' as a key in a few places which need
to understand that not every node will be considered.

The upstream logic for this node selection is in the core scheduling
packages and not readily consumable as a module. As a result, the
most reasonable path forward is to implement a small bit of the
desired functionality on our end.

For now, we just care about the RequiredDuringSchedulingIgnoredDuringExecution
field and only for node labels. This allows picking out nodes based on
labels which may satisfy most use cases.

Fixes #988

Signed-off-by: John Schnake <jschnake@vmware.com>