Add command to ouput default image registry config
This adds the subcommand `default-image-config` to `gen` to output the
default image registry config used with the `images` and `run` commands.
It will generate the config based on the version of Kubernetes that is
configured for use.

This simplifies the process for users as the registries for the E2E
tests differ for different versions of Kubernetes.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>