Tag Items with metadata about the type if item (e.g. node/file/etc)
If you look at the sonobuoy_results.yaml it can be a bit confusing
understanding the tree structure and what each node in that tree means.

This adds some extra metadata to certain Items (plugin summary, nodes,
and files) in order to make the raw file more comprehensible and
potentially aid in other processing.

Fixes #1047

Signed-off-by: John Schnake <jschnake@vmware.com>