Do not fail jobs if we can't query the API server while running
The aggregator server has a monitoring routine for each plugin. Part
of that is to poll the API server to check the pod status. Currently
job plugins will abort and be marked as failed if we have our API
query fail.

This is different than the daemonset plugins which are just blindly
tolerant of the errors.

This change makes the job plugins tolerant in the same way. This is
because we know in general that the API server was working since
the aggregator pod is running already. There are lots of transient
errors that could take place, even the API server could go down and
be brought up. None of that should cause the plugin to become a
hard failure.

Fixes #1043

Signed-off-by: John Schnake <jschnake@vmware.com>