Adds commands to print default plugin definitions
We have the `sonobuoy gen` command which prints a lot of
different types of resources and `sonobuoy gen plugin` which
prints custom plugin definitions. Sometimes it would be helpful
to be able to print the default e2e or sytemdlogs plugin definitions
so that they can be tweaked.

Added two commands `sonobuoy gen plugins e2e` and
`sonobuoy gen plugins systemd-logs` which does this.

Minor tweak to code organization to improve code reuse.

Fixes #991

Signed-off-by: John Schnake <jschnake@vmware.com>