Clarify that users shouldn't provide more options when running a file
If using the --file option for `sonobuoy run` they shouldn't also
try and set other options.

This is because we are not able to take the file output and turn
that into a genConfig object which can be merged with the other
command line options.

When in the expanded file format, we are unsure if which parts
should be merged/how.

These are meant to be exclusive options.

Fixes #1027

Signed-off-by: John Schnake <jschnake@vmware.com>