Refactor image client to not be E2E specific
The existing client for handling plugin images assumed that the images
were for the e2e plugin. We are trying to add support for other images
so we need to make the image client more generic. This change refactors
the image client to deal with image names rather than the internal
Kubernetes image `Config` type.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>