Use filepath instead of path when possible
There may be more of these situations lurking around, but as we
try to support Windows nodes we need to make sure that we are using
filepath unless we know we need forward slashes (like in URLs).

This changes a few calls to path.XYZ to filepath.XYZ and updates some
tests in the same way.

xref #732

Signed-off-by: John Schnake <jschnake@vmware.com>