Add kubernetes-version flag to images subcommands
If the cluster the user wants to get images for is not available
at the time (preparing images ahead of time, for a coworker, etc)
then the user needs a way to specify the version rather than relying
on the autodetection.

This change also causes the images from `sonobuoy images` to be printed
in a sorted order for consistency.

Fixes #1082

Signed-off-by: John Schnake <jschnake@vmware.com>