Add a new results format: manual
Manual results allow a user to write the sonobuoy-results.yaml
file largly on their own.

Due to some limitations we still wrap their file with one more
summary level item.

When the user gave us a single top-level item, the status from that
item is used for the summary. Otherwise all the top level file items
are considered and a non-judegmental routine just counts the various
status values and shows them as a human readable status for the
top level. This helps ensure that the status value from custom output
can be various pieces of info the user may want to provide and not just
something like 'pass' or 'fail'.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>