Persist images for all targets
When building the Sonobuoy images, we build images for multiple targets.
We push them in a separate step from where they were built so they need
to be persisted to the workspace. We were previously only persisting the
sonobuoy/sonobuoy tagged images, not the sonobuoy/sonobuoy-ARCH images.
This was causing CI failures as the image pushing step was trying to
push images for different architectures that did not exist locally.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>