Use passed VERSION when generating image manifest (#1099)
When calling the `push_manifest` target, we were passing `VERSION` to
use for the image tag, however it wasn't being used. Instead we using
the `IMAGE_VERSION` which was defined outside the target and was the tag
for an image that hadn't been pushed. This modifies the `gen_manifest`
target to use the passed `VERSION` value.

This also updates our Makefile to get the latest version of
manifest-tool.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>