Simplify installation instructions
Our installation instructions included a command to download and extract
a release tarball from GitHub. This command assumed the presence of
`$HOME/bin` which not all users had. It also assumed that users will be
downloading an `amd64` release.

Now that we are supporting many more platforms and architectures, it is
simpler to leave it up to the user to select the release they want and
extract it to their desired location.

Also, due to the increased changes to the platforms we support, and the
new workflows for building images for those targets, we don't recommend
building from source so remove that as a suggestion.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>