Upgrade to Kind 0.7.0 (#1110)
This upgrades our use of Kind in CI to the latest version. A recent
version of Kind introduced merging of kubeconfigs and separate
kubeconfigs are no longer written. The kind command `get
kubeconfig-path` has been deprecated so any usage of it has been
removed. The context is be updated to use the newly created cluster.
Some of our docs which reference this deprecated command have also
been updated.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>