Include custom status in results report (#1116)
Previously, results would only be included in the total number of result
items if the status for that result was one of the built-in result
statuses (passed, failed, skipped, etc). With the addition of manual
results reporting, if a user used other custom statuses, these would not
be included in the default `results` output and it would look like
Sonobuoy had processed no results.

This commit modifies the logic for the default `results` output to count
all items, not just those which match the built-in statuses. Custom
statuses with their count are also included in the output. The overall
counts for the built-in statuses are still included as before, but now
we also iterate over all other statuses and include them in the output.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>