Change default value for E2E_PARALLEL
The original implementation for enabling parallel running of tests was
added when ginkgo expected a numeric value for the number of nodes to
use when running in parallel. This is no longer needed as ginkgo can
detect the optimal number of nodes to use. The script for running the
e2e tests now expects "true", "y" or "Y" to enable parallel testing.
This commmit changes the default value to "false". Although the previous
value would not have resulted in parallel testing being enabled, this
change helps users understand that the value provided should be boolean.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>