Allow arbitrary data in Details map in results (#1130)
The previous `Details` structure in the Sonobuoy results format was a
`map[string]string`. As we are creating more plugins that are providing
reports in the Sonobuoy results format, this structure is limiting. To
allow more flexibility, the type is being changed to
`map[string]interface{}` to allow plugins to provide arbitrary data in
the `Details` field.

The version of the YAML library used when parsing the results in a
results tarball has been upgraded to v3. This is due to the fact that it
unmarshals data into a `map[string]interface{}` type when all the keys
are strings. In v2, the behaviour was to always marshal to
`map[interface{}]interface{}`. This is important as if the keys are not
strings then marshalling to JSON will fail when using the `--detailed`
mode of the `results` command.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>