Remove use of "master" in labels and names
We renamed the `master` command to `aggregator` in #847 (released in
v0.15.3). However, `master` was still being used in some of our
Kubernetes resource labels and names. We want to remove uses of
insensitive language and as such use different labels for
identifying resources.

Now, any aggregator resources will have the label
`sonobuoy-component: aggregator` and plugin resources will have
`sonobuoy-component: plugin`. The existing label is still in place as it
was used to identify the aggregator pod (which is required most sonobuoy
commands once a run has started). Removing it now would prevent users
with older clients from using newer images, or querying the state of
runs started by newer clients. Instead, both labels will be applied
howeve the `sonobuoy-component: aggregator` label is the preferred
label. If the aggregator pod cannot be found using the new label, the
client will retry using the now deprecated label. When retrying, the
client will display a warning that it is attempting to find the pod
using the deprecated label.

Newer clients will be able to query the state of runs with old labels
applied, and older clients will be able to query the state of runs
created by the newer clients.

The name of the service created by Sonobuoy has been renamed to
`sonobuoy-aggregator`. This has no user facing impact.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>