Replace MasterURL with AggregatorURL in WorkerConfig
We renamed the `master` command to `aggregator` in #847 (released in
v0.15.3). However, we left some of the existing uses in place to prevent
backwards incompatibility. This change renames the `MasterURL` field in
`WorkerConfig` to `AggregatorURL` in keeping with our terminology
elsewhere. Worker containers are configured using environment variables.
This change renames the environment variable but also adds some
compatibility for the original environment variable to still be
processed.

It is possible for different image versions to be used for the
aggregator and the worker (although it would have to be explicitly set
in the generated manifest). This changes allows older versions of
Sonobuoy to be used as the aggregator image but does not allow older
versions to be used as the worker image. Although some compatibility is
broken, this seems acceptable as we don't recommend mixing versions of
images within a single sonobuoy run.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>