Remove unused config file handling from worker
The `LoadConfig` function for the worker was configured to be able to
load configuration data from a file. However, the necessary viper
function to load and read the file was never called. This functionality
has not worked since the initial commit and so it doesn't seem
necessary. This is now being removed in favour of only configuring the
worker via environment variables.

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>