Upgrade Kubernetes client libraries to 0.18.5 (#1138)
This changes upgrades the version the Kubernetes libraries we are using
to the latest version, 0.18.5. With this, a number of the APIs had
changed to take a `context.Context` parameter. As most of the API uses
do not have a surrounding context, these updated calls take a `TODO`
context as recommended in the Golang context documentation:
https://golang.org/pkg/context/#TODO

Signed-off-by: Bridget McErlean <bmcerlean@vmware.com>