Avoid infinite loop when a following logs
When streaming logs, we previously watched a list of pods so that
we could identify when new ones were created and we would stream
their logs as well.

The problem is when the watcher breaks we do not automatically
retry the connection and a channel gets closed causing an infinite
loop which eats CPU.

This uses a RetryWatcher to solve the problem.

Fixes #1064

Signed-off-by: John Schnake <jschnake@vmware.com>