Move away from Makefile
Every dev hates the Makefile with its unique syntax and rules
and we sometimes run into issues as a result of them causing
unintended consequences. Most everyone can write/use bash and
we were really just using Make to organize/call bash functions
anyways.

Now we have a script with funtions we can source/call. If we want
we can later add argument handling so that we don't have to source
it and can instead actually call it like an executable.

Signed-off-by: John Schnake <jschnake@vmware.com>