Windows builds in CI
Adds windows builds for a number of OS versions to our CI pipeline.

Utilizes buildx and skopeo to build the windows images on the linux
runner and move them to dockerhub, then uses docker manifest to
create the manifest as usual.

In a quick reversal of a previous commit, we actually HAVE to put
all the images as "sonobuoy" and have the different arch/os use
different tags instead of different image names. This has to do with
how the manifest command looks up images and prevents issues wsuch as
"manifest not found" even though it does (as a differently named image
from the others).

Fixes #732

Signed-off-by: John Schnake <jschnake@vmware.com>