Add netapi32.dll to windows images
This missing dll is required to run may containerized golang apps
since anything checking the current user via the os package
relies on this. It so happens that klog does this and so its common
for some dependency to end up with the issue.

We can continue to shift to klog/v2 which may fix the issue, but
it would could get hit via another import as well.

Fixes #1263

Signed-off-by: John Schnake <jschnake@vmware.com>