Fix nodeSelector logic
The logic for this was previously broken, but when fixed, it
was found to actually be implemented incorrectly causing it to
overwrite the nodeSelector in all cases, not just when the podSpec
was empty.

Now it just updates the nodeSelector to linux boxes when there is
no podspec given by the user. If they want to specify a podspec to
force it to a different os (or any os) then they can do so.

Tests added to prevent this issue from regression.

Signed-off-by: John Schnake <jschnake@vmware.com>