Improve retrieve command with extraction option
Adds a flag to sonobuoy extract which allows the user
to choose to download and extract the data in a single
command. This is ideal especially when worrying about if
someone on windows 1) can untar files easily 2) can be
sure that untar will end up being aware that the tarball
uses slash pathing and the client is on Windows.

Fixes #1277

Signed-off-by: John Schnake <jschnake@vmware.com>