Adds ability for plugins to specify configmaps
The most common configuration option that we don't easily support is
a plugin specification that also includes config map info. By supporting
this we can easily support different deployment configurations without
specifying the entire `sonobuoy gen` yaml manually and sharing that.

This is particularly useful for specifying the kube-image-repo-list
as part of the plugin itself rather than as a command line option that
makes it harder to share.

Signed-off-by: John Schnake <jschnake@vmware.com>