Update the gen plugin e2e command to take a configmap
- Adds new flag to that command. It is somewhat forced in there but
it is different than the other genset and it shouldnt get all of the
gen plugin flags either IMO.
- Adds integration test for any/all commands that check for static
output. This may be gen plugin or other info that doesnt require
timestamps/UUIDs. We can extend this with a hidden field for forcing
static UUIDs or something so that we can test all gen logic this way
potentially.
- Does NOT add this to the systemd-logs plugin since there isn't really
a use case for that at this time.

Signed-off-by: John Schnake <jschnake@vmware.com>