Add sleep after systemd-logs again
The previous change which removed this sleep did also
ensure that the worker doesn't exit from errors so it doesn't
restart but the plugin container itself still exited and caused
a misleading pod status of CrashLoopBackOff.

That doesn't prevent anything for working but is confusing and
the sleep should be restored.

Fixes #1322

Signed-off-by: John Schnake <jschnake@vmware.com>