Add generic plugin transformations and use for kube-repo-list
In an effort to generalize some features to more than just e2e,
I implemented a way to tie command line flags directly to plugins
which may be loaded at a later time.

In this case, we use the transformation logic along with the new
configmap handling for plugins to add a config map and env var
to the e2e plugin whenever the user provides the flag for a custom
e2e repo configuration.

Signed-off-by: John Schnake <jschnake@vmware.com>