Default to assuming the conformance image can support progress
This was introduced in 1.17 which was numerous k8s releases ago.

We previously said we supported 3 releases back in any given Sonobuoy
release so I think it is fair to cut this out now.

It also helps in the ongoing plugin seperation changes since it
makes one less thing variable about the e2e tests specifically.

Signed-off-by: John Schnake <jschnake@vmware.com>