Rework the gen plugin commands in order to reuse more code
- Adds a new method that allows us to easily reuse all the `gen`
logic and then just pull out the plugin we want.
- Removes the ability to specify configmaps to gen plugin e2e;
this seems like a trivial loss since I just added it in order
to achieve some other goal related to the windows plugins and
it never filled much of a real need.

This PR is meant to precede another related to how the k8s-version
flags are handled. In that PR, it broke the `gen plugin e2e` command
since this command wasn't really processing the plugin in the
same way that `gen` would.

Signed-off-by: John Schnake <jschnake@vmware.com>