Apply ImagePullPolicy to all plugins not just built-in
Previously we applied imagePullPolicy to just the built in
plugins. This allowed users to specify the policy for each
plugin but made it inconvenient to change (had to manually
change the YAML) and inconsistent between our built-in
and custom plugins.

This applies the imagePullPolicy from the config to every
plugin. It occurs prior to the other plugin transformations
from flags, so if we do add something like a per-plugin
imagePullPolicy, it wouldn't interfere. I don't there will
be much demand for that though.

Signed-off-by: John Schnake <jschnake@vmware.com>