Fix docs links for 1388 issue
The docs have some special logic to try and redirect to the right
version of any given link. I've tried hardcoding the values for the
full URL but it isnt working for some reason. By just copying the file
into a few more places the links becomes much more simple.

Signed-off-by: John Schnake <jschnake@vmware.com>