Show/maintain fewer versions of docs
When we got versioned docs on the site we just left
all versions as an easy, safe choice.

Now we have so many versions that it is a bit unwieldy
when you see the dropdown and the large number of versions
makes it unnecessarily difficult to update old versions of
docs.

If users need them, they're in the git history, but otherwise
they provide little value to anyone.

Signed-off-by: John Schnake <jschnake@vmware.com>