Adjust the logic for processing plugin results for aggregator status
IF a test fails that indicates it fails, not Sonobuoy. I think the current
logic which labels the whole run as "failed" is incorrect now that the tool
has evolved to run more plugins.

Instead, the aggregator should only be marked failed if the aggregator itself
fails to get results and get them to the user.

Fixes #1414

Signed-off-by: John Schnake <jschnake@vmware.com>