Add sleep parameter to job plugins worker container
Without this, the container exits after submitting results
and leads to the pod being shut down before we can query
the logs. In some plugins (like e2e) it isn't that big of
an issue, but for plugins that dont seperately save their
logs or if we want to debug the worker container, it is
vital.

Signed-off-by: John Schnake <jschnake@vmware.com>

Worker sleep flag

FixSelflink