Removed managed fields from query output
It is a confusing waste of info and makes it much harder to look
at the output. In addition, it is not normally provided by kubectl
so users aren't even aware that it exists (I wasn't) and so it
makes the rest of the data seem more confusing than it needs to be.

Fixes #1428

Signed-off-by: John Schnake <jschnake@vmware.com>