Add flag/config setting for security context
We recently added security context to prevent policy from worrying
about Sonobuoy being a priviledged user. However, k8s for Windows nodes
does not support those settings. As a result we needed a change like this
to make it easier to un-set the default values.

I imagine we will add a few more modes here (e.g. windows) to round out
the functionality but currently we need just the default and `none` which
is the desired value currently for windows.

Fixes #1444

Signed-off-by: John Schnake <jschnake@vmware.com>