Clarify plugin list output surrounding name/file
This is a confusing point about the plugins: the filename is what
determines what sonobuoy loads from the command line, but the plugin
name (inside the yaml file) is what the aggregator server sees.

The first feedback from the blog post about this exhibted this confusion
and rightly so. I hope by adding this first line "Run as: <name>" it
will become more clear that _that_ is how you select the plugin.

Signed-off-by: John Schnake <jschnake@vmware.com>