Fix bug which failed to pass custom resultsDir
The results directory can be configured in the Sonobuoy
config, so gen needs to use that value when producing the
volume mount and env var for plugins. It was just hardcoding
different values in different places causing plugins to try
saving things in one place, and the worker loading them from
another.

Signed-off-by: John Schnake <jschnake@vmware.com>