Remove ResultsDir fields from config object
This field being configurable was an illusion; it was respected in
numerous places but not properly in the sonobuoy worker mounts or
lookups.

It was much harder/uglier to fix this and make it configurable. Since
it had been broken for years with no complaints at all, we decided to
simply remove it with a release note.

Signed-off-by: John Schnake <jschnake@vmware.com>