Sonobuoy workers report results of completed plugins
If a plugin runs to completion and exits 0 but does not write a
done file, we should still be smart enough to load up the results
directory and report those results.

This would make writing plugins much easier, especially when wrapping
tools since you can just redirect the output and the plugin is effectively
done.

Signed-off-by: John Schnake <jschnake@vmware.com>