Pass ResultsDir through necessary calls
The configurable results directory was not properly passed down,
leading to some places assuming it was configurable and other places
hard coding it.

Added code to ensure each plugin had the right path mounted as 'results'
volume mount.

Ensured the aggregator has a consistent path so that the CLI can always
know where to retrieve it from.

Signed-off-by: John Schnake <jschnake@vmware.com>