Add special config option to skip all plugins
To support the 0 plugins use case (uses sonobuoy for data
gathering only), this flag is the easiest and most clear approach.

Simply removes the plugins from the configuration before the
marshalling of the plugins. As a result, no plugins are put into
the configmap and none are available for loading on the aggregator
at run time.

Signed-off-by: John Schnake <jschnake@vmware.com>