Export field
Working on a postprocessor to handle junit files and it would
be a shame to not be able to reuse whats here already.

I think just the types and unmarshal code is needed but this may extend
a bit if necessary to avoid too much copy/paste.

Field needed to be exported as well.
Signed-off-by: John Schnake <jschnake@vmware.com>