Fix template for namespaced permissions
- do not create namespace
- namespace should be in metadata, not metadata.label
- in clusterAdmin mode, no need for namespace in metadata.label
- do not use non-resource URLs for namespaced role as they are not allowed

I think the template is too hard to read at this point and we should
translate it into code so it is easier to read, modify, and debug.

There are other issues with the run code itself that will cause the namespaced
permissions to fail, but this is the first step.

Signed-off-by: John Schnake <jschnake@vmware.com>